Implement container attribute parsing for `custom`
See the module docs. Basically, we allow bypassing the built-in parsing to customize one or more attributes. If you're going to customize all of them, then there's no point in using the derive macro, however I found there are cases where all but one attribute can be derived, and it stinks to resort to needing to manually re-implement when that happens.