Add `custom = <function>` container attribute and allow optional `ignore = "<reason>"` for fields (#6)
* Implement container attribute parsing for `custom`

See the module docs. Basically, we allow bypassing the built-in parsing to customize one or more attributes. If you're going to customize all of them, then there's no point in using the derive macro, however I found there are cases where all but one attribute can be derived, and it stinks to resort to needing to manually re-implement when that happens.

* Allow `cache_diff(ignore = "<reason>")` && checks for "custom"

Allows someone to leave behind a comment for why something is ignored as well as if someone uses `ignore = "custom"` we'll enforce that the struct must have a `cache_diff(custom = <function>)` defined on the struct.

* Changelog + README

* Update version

Trying to get version resolution with the build packs-ruby crate to work with manually specified git

```
cache_diff = { features = ["bullet_stream"], branch = "schneems/custom_fn", git = "https://github.com/heroku-buildpacks/cache_diff" }
```

Not sure if it will work.

* Update version

* Fix accidentally parsing other container attributes

Even though the derive macro knows which attributes it owns:

```
#[proc_macro_derive(CacheDiff, attributes(cache_diff))]
pub fn cache_diff(item: TokenStream) -> TokenStream {
    //...
}
```

It passes in the full AST. The non-cache-diff attributes must be filtered out. This is how the `CacheDiffAttributes` (which holds Field attributes) is implemented.