Refactor + Add unit tests (#8)
Close #7. This major refactor takes the approach of attempting to make invalid state impossible to represent. It's not 100% there but it's much better.

At a high level we have two primary ways of configuring CacheDiff: container attributes (attributes on the struct e.g. `struct Metadata {...}` ) and field attributes (e.g. `name: Version`). We need to parse these attributes, to do that I made a pattern where we use strum to build an enum that represents all possible attribute states, this is a `ParsedAttribute`, we have strum generate a set of known valid keys and use this to parse our attributes so the code will never go out of sync with the API.

Once all attributes are parsed they're used to build up state while checking for invalid configuration options.

For fields we build ParsedAttribute -> ParsedField -> ActiveField.

For containers we build ParsedAttribute -> <parse fields> -> CacheDiffContainer.

This pattern also makes it easier to add unit tests which I've done.