Check all keys are utilized
The current type pattern will guarantee that every attribute known to the system is parsable, but it cannot guarantee that it's ever pulled from the HashMap. There's a slim chance someone could add an attribute, but forget to look it up. 

It's not likely with our current test suite, but I'm exploring ways to bake in more safety on edge cases where I can't lean more on types.

This code will check to ensure that all entries are removed from the attribute lookup, so if one is parsed but not used, they'll get a nice error message.

Worth noting: I experimented with a "drop guard" style of check that would automatically perform this logic on drop, but that leaves me with only the ability to panic in the error scenario which doesn't give us nice formatted compile errors.